create ListElements clas
